+++ src/libs/TransactionRecoveryApi/index.ts
@@ -2,3 +2,4 @@ export * from "./AxelarDepositRecoveryAPI";
 export * from "./AxelarGMPRecoveryAPI";
 export * from "./constants/error";
 export * from "./AxelarRecoveryApi";
+export * from "./AxelarTransferAPI";