+++ src/libs/test/AxelarQueryAPI.spec.ts
@@ -67,7 +67,7 @@ describe("AxelarQueryAPI", () => {
   });
 
   describe("estimateGasFee", () => {
-    test("It should return estimated gas amount that makes sense for USDC", async () => {
+    xtest("It should return estimated gas amount that makes sense for USDC", async () => {
       const gasAmount = await api.estimateGasFee(
         EvmChain.AVALANCHE,
         EvmChain.ETHEREUM,
@@ -125,6 +125,7 @@ describe("AxelarQueryAPI", () => {
         tokenAddress: "uaxl",
         decimals: 6,
         common_key: "uaxl",
+        mintLimit: 0,
       });
     });
   });