+++ src/libs/types/index.ts
@@ -75,6 +75,12 @@ export interface BaseFeeResponse {
   success: boolean;
   error?: string;
   baseFee?: string;
+  sourceToken?: {
+    gas_price: string;
+    decimals: number;
+    name: string;
+    symbol: string;
+  };
 }
 
 export type CosmosBasedWalletDetails = {
@@ -103,6 +109,10 @@ export interface AxelarSigningClientConfig extends AxelarQueryClientConfig {
 }
 
 export type AxelarRecoveryAPIConfig = AxelarQueryAPIConfig;
+
+export type AxelarTransferAPIConfig = {
+  environment: Environment;
+};
 
 // Includes all native tokens and stablecoins
 export enum GasToken {
@@ -154,6 +164,38 @@ export interface QueryGasFeeOptions {
   estimatedGas?: number;
 }
 
+export interface QueryTransferOptions {
+  depositAddress?: string;
+  recipientAddress?: string;
+}
+
+export enum QueryTransferStatus {
+  DEPOSIT_CONFIRMED = "deposit_confirmed",
+  ASSET_SENT = "asset_sent",
+  VOTED = "voted",
+  BATCH_SIGNED = "batch_signed",
+  IBC_SENT = "ibc_sent",
+  EXECUTED = "executed",
+}
+
+export interface QueryTransferResponse {
+  success: boolean;
+  error?: string;
+  data?: {
+    id: string;
+    status: QueryTransferStatus;
+    type: string;
+    amount: number;
+    fee: number;
+    denom: string;
+    senderChain: string;
+    senderAddress: string;
+    recipientChain: string;
+    recipientAddress: string;
+    blockExplorerUrl: string;
+    blockHeight: number;
+  };
+}
 export interface AxelarTxResponse extends DeliverTxResponse {
   rawLog: any;
 }